Move crypto message test out of the realtime suite, making it a pure unit test.
